- Modularize sandbox project files
- Enhance Computation Analysis box execution time
