+++ sandbox/extensions.py
@@ -1,0 +1,4 @@
+from flask_caching import Cache
+
+
+cache = Cache()
