+++ sandbox/requirements.txt
@@ -2,3 +2,5 @@ requests==2.31.0
 pyarrow==15.0.0
 pillow==10.2.0
 rasterio==1.3.9
+python-dotenv==1.0.1
+Flask-Caching==2.3.0  # https://flask-caching.readthedocs.io/en/latest/
